+++ src/components/PageSidebar/PageSidebar.tsx
@@ -1,108 +1,20 @@
-import React, { useState } from "react";
-import { Link } from "react-router-dom";
+import React from "react";
+import { styled } from "@mui/system";
+import TopSidebar from "./TopSidebar";
+import BottomSidebar from "./BottomSidebar";
 
-import { styled } from "@mui/material/styles";
-
-import Box from "@mui/material/Box";
-import Drawer from "@mui/material/Drawer";
-import List from "@mui/material/List";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-
-import { ReactComponent as Logo } from "src/images/svg/Logo.svg";
-
-import { sidebarArray } from "./SidebarInterface";
-
-const CustomDrawer = styled(Drawer)({
-  "& .MuiPaper-root": {
-    backgroundColor: "#01070E",
-  },
-  "& .MuiBox-root": {
-    backgroundColor: "#051424",
-    color: "#869AAF",
-    "&:hover": {
-      color: "#FFFFFF",
-    },
-  },
-  "& .MuiBackdrop-root": {
-    backgroundColor: "transparent",
-  },
+const NormalSidebar = styled("div")({
+  width: "10%",
+  height: "100%",
+  borderRight: "2px solid #030f1c",
 });
-const CustomBox = styled(Box)({
-  "& .MuiListItemText-root": {
-    padding: 0,
-    color: "#869AAF",
-    "&:hover": {
-      color: "#FFFFFF",
-    },
-  },
-  "& .MuiListItemIcon-root": {
-    padding: 0,
-    margin: 4,
-    minWidth: 25,
-  },
-  "& .MuiListItem-root": {
-    width: "100%",
-    "&:hover": {
-      backgroundColor: "#0247FF",
-      color: "#FFFFFF",
-    },
-  },
-});
-
-const styles = {
-  color: "#FFFFFF",
-  textDecoration: "none",
-};
-
-const PageSideBar = () => {
+const Sidebar = () => {
   return (
-    <div>
-      {(["left"] as const).map((anchor) => (
-        <React.Fragment key={anchor}>
-          <CustomDrawer variant="persistent" anchor={anchor} open={true}>
-            <CustomBox
-              sx={{
-                width: 170,
-                paddingTop: 2,
-                "@media (max-width:1100px)": {
-                  width: 50,
-                },
-              }}
-              role="presentation"
-            >
-              <ListItemIcon
-                sx={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                  width: 150,
-                }}
-              >
-                <Logo />
-              </ListItemIcon>
-              <Box sx={{ marginTop: 4 }}>
-                {sidebarArray.map((item) => (
-                  <List key={item.text}>
-                    <ListItem key={anchor} disablePadding>
-                      <Link style={styles} to={item.path}>
-                        <ListItemButton>
-                          <ListItemIcon>{item.icon}</ListItemIcon>
-                          <ListItemText> {item.text}</ListItemText>
-                        </ListItemButton>
-                      </Link>
-                    </ListItem>
-                  </List>
-                ))}
-              </Box>
-            </CustomBox>
-          </CustomDrawer>
-        </React.Fragment>
-      ))}
-    </div>
+    <NormalSidebar>
+      <TopSidebar></TopSidebar>
+      <BottomSidebar></BottomSidebar>
+    </NormalSidebar>
   );
 };
 
-export default PageSideBar;
+export default Sidebar;
