+++ src/components/UI/commonStyles.tsx
@@ -31,5 +31,5 @@ export const ViewButton = styled("button")({
 export const ViewAll = styled("div")({
   display: "flex",
   justifyContent: "center",
-  marginTop: "30px",
+  marginTop: "20px",
 });
